feat: databend-meta watch API provides `initial_flush` field
If `WatchRequest.initial_flush` is `true`, when the watch stream is
established, all the key values in the specified range will be sent
through the stream to the watching client.

This allows the client to setup a full copy of the key range as a client
side cache, and furthur modifications will be received and the client
side cache can be updated.