fix(meta-semaphore): deadlock between meta-service semaphores (#17986)
### Problem

Before this commit, when a smaller seq semaphore is inserted after a
larger seq one, the smaller one may not be acquired because the capacity
is exhausted. This might lead to a deadlock:

For example, `A` has seq 1, `B` has seq 2. the capacity is 1.
- Process `A` received sem `B` first, then sem `A`;
- Process `B` received sem `A` first, then sem `B`.

```text
|             acquired    waiting
| Process A:  [ B ]       [ A ]
| Process B:  [ A ]       [ B ]
```

And then two processes blocks each other:
- Process `A` await sem `B` to be removed from `acquired` queue, which await Process `B` to remove sem `B`.
- Process `B` await sem `A` to be removed from `acquired` queue, which await Process `A` to remove sem `A`.

### Solution

In this commit, a semaphore permit is always acquired even when the
capacity is full, if it's seq is smaller than the last one in the
`acquired` queue, which means this smaller seq is inserted out of order.

Therefore a smaller seq semaphore can always be acquired on each
databend-query node. There won't be a deadlock.

Tons of thanks to @dqhl76 for addressing this issue!