fix(meta-semaphore): skip dup seq semaphore permit (#18003)
# Problem

Up to 1.2.742 datbend-meta might send a kv change event befoore
initialization events in a watch stream, which is a bug.

This cause the meta client side semaphore receives duplicated `add` kv
event, such as two `key=foo, None -> Some(val)` events. Such events
sequence inserts duplicated semaphore permit in the acquired queue and
waiting queue. And a `remove` kv event just remove one of the two
permit, which leads to a never permit that can never be released.

# Solution

In this commit, when a duplciated seq permit is received, just skip it,
instead of insert it into `waiting` queue.