fix(meta-service): when setting up a watch stream, lock state machine (#18005)
Before this commit, a watch stream is built in two steps:
- add a watch sender to the dispatcher,
- then send a bunch of kvs for initialization.

But there is a gap between these two steps. For example, a key might be
inserted between these two steps. In such a case, the client will receive
two `add` events, such as: `key=foo, None -> Some(val)`, one of them is
emitted when `key=foo` is added, the other is emitted by the
initialization process.

These two identical events will confuse the client in some use cases,
such as the meta-based semaphore does not expect two identical
events to be received.

In this commit, the two steps of setting up a watch stream are done
atomically, with the state machine lock held. So that the client won't
receive duplicated events.