refactor(metactl-import): when importing, also write the data version
Before this commit, when importing, the data version (such as `V004`)
is not written to disk, and the data version is assumed to be the
oldest compatible version, and the on-disk data then will be upgraded to the
latest version. For example, when `databend-metactl` imports `V004`
data, the data on disk is left as `V002`, and then the data is upgraded
from `V002` to `V003` and finally to `V004`.

This is an unnecessary burden, and the upgrading process invokes the legacy
`sled-db` based storage, which leads to problems on NFS (`device busy`
error when deleting sled from disk after closing it).

This commit refines this behavior by directly writing the data
version such as `V004` to disk after importing. Thus there won't be an
unnecessary burden of upgrading the data, and it gets rid of the
necessity to access sled DB, which is a workaround for running on NFS.