refactor(meta-client): remove usage of obsolete `prev_value`
`TxnPutRequest.prev_value` and `TxnDeleteRequest.prev_value` inform the
server to respond with the value before the put or delete request.

Since 1.2.304, the meta-server always responds with the prev-value and
these two fields are never used.

The lease supported meta-server version is now 1.2.677, therefore we can
completely remove these two fields from the protocol.

This change is non-breaking.