fix(query): add_months need to respect last day of month
ADD_MONTHS returns slightly different results than DATEADD used with a MONTH component:

For both ADD_MONTHS and DATEADD, if the result month has fewer days than the original day, the result day of the month is the last day of the result month.

For ADD_MONTHS only, if the original day is the last day of the month, the result day of month will be the last day of the result month.

In main:

```sql
select add_month('2025-04-30',1)
----
2025-05-30
```

In PR:

```sql
select add_month('2025-04-30',1)
----
2025-05-31
```