fix: meta-semaphore: re-connect when no event recevied
Refactor permit implementation to properly handle acquirer closure and
connection errors.

The permit now uses oneshot channels to detect when the underlying
permit entry is removed from meta-service, replacing the previous
BoxFuture. Because a future that is not polled will block the channel.

If the meta-service does not send an event for `permit_ttl * 1.5`,  the
connection will be re-established.

Add `AcquirerClosed` error type for client side close error.