fix(meta-service): detach the SysData to avoid race condition (#18722)
* fix(meta-service): detach the SysData to avoid race condition

When creating a new level in state-machine, it should detach the SysData
to avoid race condition with snapshot building.

Before this commit, the new writable level and the snapshot compactor
shares the same data thus the new applied data increases the
`last-log-id` of a new built snapshot. Result in a snapshot that
lacks some log entries it declares to have.

* M  src/meta/raft-store/src/sm_v003/compact_immutable_levels_test.rs