fix: vacuum tables dropped by `create or replace` statement
Tables implicitly dropped by the `create or replace` statement are not handled correctly:

The vacuum operation collects candidate tables according to the `drop_on` mark of table meta data,
unfortunately, during `create or replace`, the table being replaced is NOT marked with `drop_on`.

In this PR, for table ids that have no `drop_on` marks, will be selected as the candidates for
vacuum even, except it is the largest table id (of the same table name), which is the live table that
should be kept.